refactor(llm): consolidate Anthropic adaptive-thinking gating into shared version parser
opencode_config.py kept its own frozenset of adaptive-thinking models while
multi_llm.py parsed the model version — two sources of truth that had already
drifted: the frozenset was never updated for the Claude 5 line, so Craft
configured claude-fable-5 et al. with the legacy enabled+budgetTokens config
those models reject.

Move the version parser into onyx/llm/anthropic_versions.py with two named
gates matching the two thresholds the call sites actually encode:
- supports adaptive (4.6+): used by opencode_config; replaces the frozenset
- requires adaptive / rejects sampling params (Opus 4.7+): used by multi_llm

Chat-path behavior is unchanged. Craft configs change for three cases, all
fixes: Claude 5 models now get adaptive thinking, claude-opus-4-6 regains it
(dropped incidentally in #11669), and vendor-prefixed/suffixed ids like
anthropic.claude-opus-4-7-20260115-v1:0 now parse.
